requirements/app.txt
django==4.0.1
djangorestframework==3.13.1
psycopg2==2.9.3
django-environ==0.8.1
drf-yasg==1.20.0
